Move package discovery from project source
